Exclude subpartitions for given root partition in leaf-partition-data mode
gpbackup with the --leaf-partition-data and --exclude <root_partition> options
adds leaf partitions to the recovery plan, but the root partition is not
included in the schema creation script and gprestore fails with an error.

The solution is to exclude all leaf partitions for a given root partition
in leaf-partition-data mode.