Fix gprestore with --resize-cluster to a smaller one
When restoring a large cluster to a smaller one with the --resize-cluster
option, the error "Found incorrect number of backup files" occurs.
This error occurs when checking a backup, when the pgrestore expects
an increased set of data in the backup directories for certain segments,
but in fact this is not the case. Moreover, if you disable the checking,
another error will appear indicating that not all expected data was restored.

This patch fixes both of these bugs.
To correct the first error, the patch collects actual data not only from
the directory for one segment, but also from the remaining directories
that will be restored on this segment.
To correct the second error, the patch changes the path to the file
being restored, replacing not only the file name,
but also the directory name. This can be done because the path always
contains the sequence "...SegmentNumber/backups/..."

Tests for restoring with the --resize-cluster option to a smaller cluster
contained the moveSegmentBackupFiles function, which moved the backup data
to the required directories. Now you don't need to do this anymore.