gpbackup incorrectly backups HyperLogLog statistics
Sometimes gprestore cannot restore such backup.
This happens if HyperLogLog statistics were collected for a column whose
type cannot be implicitly converted to text, for example INTEGER.
In this case, gpbackup backups the binary representation of such statistics,
converting them to a column type, which is incorrect, since HyperLogLog
statistics are stored in a binary representation of the BYTEA type.

This patch corrects the described issue by explicitly
specifying the BYTEA type for HyperLogLog statistics.

In the test, the leaf partition was excluded from checking
the number of statistics, because the current version of gpbackup
does not backup statistics for leaf partitions.
Once this is corrected, this additional condition can be removed.