Implement extension dependency for gpbackup (#46)
gpbackup does not take into account extension dependencies and places commands
to create extensions in a random order. This can lead to restoring errors if an
extension is created that depends on another that has not yet been created.

This patch adds recursive sorting of extensions by dependencies.
For 5X, simple sorting by OID is used, because there is no recursion support.