Fix backup statistics for leaf partitions (#42)
gpbackup does not backup statistics for leaf partitions with
leaf-partition-data option

This patch solves the problem by backuping statistics for all specified
tables, including leaf partitions, and not just root partitions.

The test for this patch is based on the test for the 85384fa.
And, as stated in that patch, it removes the unnecessary
condition when comparing the number of statistics.

Cherry-picked-from: 0ce7382
(cherry picked from commit 6b5fe21d6d592ab37d554ad5593308aebadf61c6)