The gprestore helper skips distributed tables if there are replicated ones. (#48)
That leads to freeze of the restore process.
This occurs due to an incorrect condition when restoring to a smaller cluster.
I corrected this condition and also extended the test to restore replicated
tables on a smaller cluster so that it checks for the identified error.

(cherry picked from commit 89fe31635dfdecd5f7d5879f46e985522f69db2f)
Code was removed, because upstream reworked solution