Fix backup table ordering tests
Upstream commit 7833212 removed the ExpandIncludesForPartitions function,
but it continued to be used in our tests in commit 1ddc17e. Because of this,
integration tests could not run and would throw compilation errors.

Failed to compile integration:
./wrappers_test.go:168:16: opts.ExpandIncludesForPartitions undefined (type *options.Options has no field or method ExpandIncludesForPartitions)
./wrappers_test.go:184:16: opts.ExpandIncludesForPartitions undefined (type *options.Options has no field or method ExpandIncludesForPartitions)

I corrected our tests so that there were no compilation errors
and they passed successfully.