gpbackup does not use shared snapshot for first table in worker processes
gpbackup:

1) creates a connection pool
2) opens transactions for each connection
3) takes and exports a shared snapshot of the first connection
4) spawns worker processes for all connections (except the first one),
   which perform the following actions:
    a) if the transaction is not yet open, then it is opened and
       the shared snapshot from step 3 is imported into it
    b) the table is copied using the COPY command
    c) if there is a shared snapshot, then the transaction is closed
    d) go to step 4.a

As a result, the first time step 4.a is skipped when copying the first
table in the worker process, because the transaction was opened in step 2.

The solution is to close the open transaction (after step 4 and before step
4.a) to open a new transaction with the shared snapshot below (at step 4.a).

A check has been added to the tests to ensure that the first table is also
processed in the shared snapshot.