Implement gpdb 7 support for gpbackup ordering patch
gpbackup used pg_partition and pg_partition_rule to select partitions for gpdb6
and prior, but gpdb7 ditched them. This patch adds a new query for gpdb7 which
uses pg_partition_root and pg_inherits. Also, gpdb7 complains on the query from
GetExtension saying that it's ambiguous. This patch also fixes it