Rework patch that excludes leaf partitions by the root (#60)
Previous patch 1be2727 conflicts with new patch 7833212

I reworked it. The new patch 7833212 changed the logic of the
getOidsFromRelationList function (on which I based my
getExcludedRelationOidsList function in the previous patch 1be2727) from an SQL
query to using pre-populated IncludedRelationFqns and ExcludedRelationFqns maps.
Therefore, I also had to use these maps instead of an SQL query. I slightly
expanded the structure of the map by adding the fields RootOid - for the Oid of
the root partition and Name - for the name of the current table. Based on this
data, in the ValidateTablesExist function, if necessary, I add the necessary
relationships to the ExcludedRelationFqns exclusion map.

gpbackup with the --leaf-partition-data and --exclude <root_partition> options
adds leaf partitions to the recovery plan, but the root partition is not
included in the schema creation script and gprestore fails with an error.

The solution is to exclude all non-root partitions for a given root partition in
leaf-partition-data mode.

The first test checks that if root partition exclusion and the
leaf-partition-data option are specified, then non-root partitions should not be
included in the restore.

The second test checks that if leaf partition exclusion and the
leaf-partition-data option are specified, then the patch does not change the
behavior: only this leaf partition is excluded from the backup. The behavior of
the second test for 6X and 7X is different: for 6X the leaf partition is
restored empty, and for 7X it is not restored at all.

The third test checks that if only a root partition exclusion is specified, then
the patch does not change the behavior: the table does not include into the
backup.

The integration tests "creates a one-level partition table", "creates a
two-level partition table", "creates a GPDB 7+ root table" and "correctly maps
oids to parent or leaf table types" have been changed. In the first three, I
added the initialization of the new Name field of the PartitionLevelInfo
structure. And for the latter, I also added the initialization of the new
RootOid field. Real values from the database are expected for comparison.