Clean up segments helper processes after error during restore (#62)
The data.restoreDataFromTimestamp function launches helper processes on segments
during a single data file restore or during a resize restore. At the same time,
the restore.DoCleanup function cleans these processes only in a single data file
restore, and does not clean them in a resize restore. This can result in helper
processes on segments being abandoned after error during restore. This patch
solves this problem by causing cleanup in the second condition as well. For
testing, a test backup was prepared, which is restored with an error.