Handle tables whose parent is in the extension
In 7X, if you add a partition to the partitioned table created by the extension,
the backup attempt would fail. The reason was that when receiving information
about table inheritance, tables related to extensions were filtered out, and
therefore the added partitions did not have a link to the parent partition.

This patch fixes the getUserTableRelations function to ignore partitions whose
root is in the extension, as is done in 6X. The filtering of extension tables
from GetTableInheritance has also been removed in order to correctly handle
situations when the parent table is added to extensions.