Add the ability to backup partitions whose parent is in the extension
This patch adds the ability to backup partitions of tables in the extension that
were added after the extension was created.

Several issues have also been fixed:
1. We no longer get panic when trying to backup a 7X cluster in which there are
partitioned tables in the extension, some partitions of which are outside the
extension. This was due to the fact that when we received the parent tables, we
filtered out the tables in the extension.
2. In 6X, if the table was inherited from a table in the extension, we we no
longer lost inheritance during backup.
3. When backing up a 6X cluster, we no longer ignore partitions not from the
extension of tables in the extension, but only if the '--leaf-partition-data'
flag is present, since the COPY command to the intermediate table in 6X does not
return anything.