Make it possible to backup/restore extension configuration tables (#75)
This patch adds support for extension configuration tables to gpbackup. The
extension script can specify tables that need to be backuped/restored using the
standard function pg_catalog.pg_extension_config_dump. As a result, gpbackup
will include the data of the specified tables (but not their definition) in the
backup. If the second argument to the pg_catalog.pg_extension_config_dump
function is an empty string, then gpbackup will dump the entire contents of the
table. A non-empty string in the second argument of the
pg_catalog.pg_extension_config_dump function means a WHERE clause that filters
out the data to be dumped. GPDB versions earlier than 7 do not support the
IGNORE EXTERNAL PARTITIONS option for the COPY ( query ) TO command, so this
option is omitted if an additional filter is present.

(cherry picked from commit e75a86aa7c785f35f6f24cdc903828b8b481bba8)