Fix gprestore hang on errors with on-error-continue option when resize-restore (#76)
Fix gprestore hang on errors with on-error-continue option when resize-restore

gprestore did not correctly handle the on-error-continue option when
resize-restore. Errors during restore could cause restore helpers to freeze.

This patch adds the creation of skip files in case of errors, as well as error
checking on segments during resize-restore, as is already done with
single-data-file.