Fix gprestore handle copy error with jobs > 1 (#78)
Fix gprestore handle copy error with jobs > 1.

When using the jobs > 1 option, if a copy error occurred, the corresponding
goroutine was terminated. But other goroutines continued to wait for COPY and
restore froze.

This patch adds the use of execution context to goroutines, which allows
requests to be canceled on errors.