Fix restoring with subset and --on-error-continue together
Subset restore API is incompatibilite with enabled "--on-error-continue" flag.
When some error is occured at reading data from file, plugin does not know about
such error and continues sending data to the pipe. It leads to situation:
gprestore breaks reading data of the current relation (for example because relation
does not exist) and starts reading data of the next relation (because option
--on-error-continue is set), but plugin sends data of non-existing relation.

This patch disables subsets, if the "--on-error-continue" option is set. Also
unit tests were added for function, which checks if it allow to use subset or
not.

Co-authored-by: Dennis Kovalenko <kds@arenadata.io>