Add the ability to backup partitions whose parent is in the extension
This patch adds the ability to backup partitions of partitioned tables whose
root is in the extension for gpdb 7.
We can backup partitions that are outside the extension if no data is added to
the default partition in the extension creation script, otherwise we will get an
error when restoring from the backup. Because when adding a new partition, the
constants on the default partition change, and if the data in the partition does
not meet the new constraints, we will get an error. If such a situation occurs
during restore, you should exclude problematic partitions using --exclude-table,
where you need to specify the name of the root partition.

Also fixed panic in 7X in case parent tables are in the extension.