Use separate error buffers for different processes in gprestore helper
Problem description:
gprestore helper reused one error buffer for multiple processes when it worked
with a plugin. It resulted in the following problem: if the plugin was launched
several times, the error was detected only on the 2nd run, while the error
happens on each plugin run.

Fix:
Add a separate error buffer for each RestoreReader.

For the backup helper case, keep one error buffer as only one backup plugin
instance is started for each backup helper (but now the buffer is made local in
doBackupAgent).