Backup statistics by one row (#77)
gpbackup backuped statistics, loading them entirely into RAM. This can create
difficulties in the case of a huge number of tables and may come down to OOM.

This patch changes the backup logic: now statistics are not loaded entirely at
once, but are requested in portions by one row. This significantly reduces RAM
consumption when backing up statistics.

The test was not added because the result does not change. And to watch the
difference in memory consumption, we need a lot of tables with large statistics,
which is not very good for running tests.