Add the ability to backup partitions whose parent is in the extension (#74)
In 7X, if there was a partition in the database whose parent was in the
extension, then the backup attempt would have ended in panic. This was due to
the fact that when receiving the names of the parent tables, the tables in the
extension were filtered out, which led to a reference to nil when creating DDL
for ATTACH PARTITION.

This patch fixes the panic by removing the filtering of tables in the extension
for 7X.

This patch also adds the ability to backup such partitions in 7X. For 6X and
earlier, the old behavior was left in order not to violate the established
behavior.

There is a situation when, after a backup of the database with such partitions,
an error may appear when restoring. This can happen if data is added to the
default partition in the extension creation script. Then, during the restore,
when creating a new partition, the default partition constraints will change,
and if the data in the default partition does not meet the new constraints, an
error will appear. If such a situation occurs, problematic partitions should be
excluded using --exclude-table, where a root partition is specified. There is a
test for this case: "allows the user to exclude partitions that hinder restore
if the default partition in the extension contains data".