Fix gprestore calculates single data file offsets on errors (#91)
If, when restoring a backup single-data-file, an error occurred during copying,
then gprestore calculated the last offset by adding the copied number of bytes.
If at the next iteration there was no error in the same place, then this last
offset was replaced by the end. In the event of a repeated error, this did not
happen and, as a result, the last offset was calculated incorrectly, which led
to data being read at incorrect offsets, errors and freezes.

This patch fixes the calculation of the last offset in case of errors.

(cherry picked from commit 167da8eb2e6c7c88b583c9e7185ed560b33b07fd)