Store type for statistic as fqdn. (#104)
Store type for statistic as fqdn.

This patch adds storing type for exported statistics as FQDN. Previously if
different schema was used for custom type definition, statistics could not be
restored because type name was used with no explicit schema definition. This
patch adds proper namespace to the type when backup generating restore
statements.