Make signal handler test more stable
Sometimes tests of the signal handler can fail due to a long start of
gpbackup/gprestore. This can happen due to slow storage and a cold disk cache.
In this case, the process can receive a signal even before the signal handler is
initialized.

This patch adds waiting for the first line in stdout before starting the timer
before sending a signal.