Fix silent crash of helper with plugin during resize-restore (#120)
Commit ae78091 added skipping batch processing when a skip file was detected.
But in one place it was done incorrectly, which led to the use of uninitialized
readers while waiting for the plugin to close. In this case, the helper
silently exited without creating an error file and without writing anything to
the log. As a result, the main process hung for 300 seconds on each remaining
table until the very end, without restoring them.

To reach the crash site, the resizing factor must be greater than two.

Since the skip file is detected after the plugin has started, we need to wait
for the plugin to finish working in this case as well. But after the skip file
is detected, plugins will no longer start and readers will not be created.
Therefore, the solution to the problem is to simply check the readers for
initialization.

Ticket: ADBDEV-6750