Add test for doRestoreAgent in gpbackup (#102)
Add tests for doRestoreAgent with interface mocking.

This PR proposes to add mocking interfaces and unit test for the
doRestoreAgent() function in the gprestore.

When we look at the gprestore code, and into doRestoreAgent() in particular,
there is not so trivial logic which might be difficult to cover with the
end-to-end and integration tests. From the other side this logic relies on
numerous external dependencies which makes unit testing nearly impossible.

This commit proposes using of well-known mocking technique for unit testing of
gpbackup code with an example based on gprestore and doRestoreAgent(). With
interface mocking it becomes possible to:

    extend test coverage with more specific use cases
    make testing closer to the golang code, resulting in fast tests with no
    external dependencies, tests will be more stable
    test is using a single process which makes it faster and easier to debug

However, this approach only complements existing end-to-end and integration
tests, it's not intended to replace them. The outcome heavily relies on the
clarity and quality of mock interfaces; it's easy to achieve unintended results.

The idea for mocking is to replace complex objects with their test imitation.
Unfortunately, gpbackup is written in solid manner, and it's not possible to
just replace built-in functions. To mitigate it, internal dependencies are split
with introducing of interfaces for dependencies external related to code under
test. Production code is using production interface implementation, while in
testing code we use test implementation, which does do only things related to
the test and return values exactly for the test scenario.

This commit adds the following interfaces:

    IRestoreHelper - an abstraction for restore helper functions located in
    helper/restore_helper.go
    IRestoreReader - an abstraction for file reader used for restoring data
    IPluginCmd - an abstraction needed to keep track of readable stderr and
    whether the command has already been ended

This is a minimum change for the supplied tests. It's expected that when more
methods or interfaces will be needed they will be created of necessity. The
changes for the interface introduction are minimal and straightforward.

Test logic is located in the helper/helper_test.go.