Fix pre-creating the required number of pipes in gprestore (#125)
Commit 2142387 introduced a problem with pre-creating pipes. gprestore expected
a pipe to be pre-created for each job/copy-queue-size. But the existing logic
was pre-creating pipes for batches of tables instead of the tables themselves.
The main process starts several goroutines, for some of which it does not
pre-create pipes, which leads to a 300-second wait for the pipe to be created
and a subsequent error in restoring the corresponding table. This patch solves
the problem by pre-creating the required number of pipes. The required number
is equal to the multiplication of the number of connections (which is equal to
jobs or copy-queue-size) and the number of batches, because the main process
starts goroutines based on the number of connections.

The tests were not added because it would require waiting in the tests for more
than 300 seconds, which is not very nice.

Ticket: ADBDEV-6893