Update gpbackup to support external Orca (#126)
Problem description:
gpbackup tests require the 'dummy_seclabel' Greenplum contrib module. But, when
setting this module to 'shared_preload_libraries', previous value of 
'shared_preload_libraries' is cleared out. Therefore, in case of GPDB with
external Orca (when Orca is a shared library), Orca is disabled during testing.
As Orca is an essential part of GPDB, testing the gpbackup with disabled Orca is
not correct.

Fix:
In setting of 'shared_preload_libraries', read current value, and add or remove
only 'dummy_seclabel' substring.
Plus, README file is updated with a proper command sample for 'gpconfig'.