Cleanup in gprestore kills only own helpers (#127)
If several restores of the same backup are launched simultaneously, then the
very first restore that completes during cleanup kills not only its helpers,
but also all helpers working with the same backup, because the processes are
filtered only by same TOC-file. The patch solves the problem by adding a filter
by the name of the pipe, which already has the PID of the main process of the
restore. gpbackup is not affected by this problem, because it will always have
a unique TOC-file.

Ticket: ADBDEV-7175