Migrate gp-common-go-libs to GreengageDB (#128)
This changes the dependency URL for gp-common-go-libs. Changes are required
for all .go files that import it, and also for go.mod and go.sum. Go
version and the libraries' versions had to be updated since Greengage's
gp-common-go-libs is a more fresh version than what we've been using before,
so it requires newer versions.

Ticket: ADBDEV-7077